+++ src/main/java/dev/nullkeeper/strengthsmptrollitems/ravager/RavagerDefaultVisibility.java
@@ -1,0 +1,12 @@
+package dev.nullkeeper.strengthsmptrollitems.ravager;
+
+import org.bukkit.entity.Ravager;
+
+@FunctionalInterface
+interface RavagerDefaultVisibility {
+    void hide(Ravager ravager);
+
+    static RavagerDefaultVisibility paper() {
+        return ravager -> ravager.setVisibleByDefault(false);
+    }
+}